refactor(api): move logic from controller to service
